# A couple logging performance improvements + Benchmarks