--- a/src/WebJobs.Script/Properties/AssemblyInfo.cs
+++ b/src/WebJobs.Script/Properties/AssemblyInfo.cs
@@ -16,5 +16,6 @@ using System.Runtime.InteropServices;
 [assembly: InternalsVisibleTo("Microsoft.Azure.WebJobs.Script.WebHost")]
 [assembly: InternalsVisibleTo("Microsoft.Azure.WebJobs.Script.WebHost.Core")]
 [assembly: InternalsVisibleTo("Microsoft.Azure.WebJobs.Script.Tests")]
+[assembly: InternalsVisibleTo("Microsoft.Azure.WebJobs.Script.Tests.Benchmarks")]
 [assembly: InternalsVisibleTo("Microsoft.Azure.WebJobs.Script.Tests.Integration")]
 [assembly: InternalsVisibleTo("DynamicProxyGenAssembly2")]