--- a/test/Benchmarks/Microsoft.Azure.WebJobs.Script.Tests.Benchmarks.csproj
+++ b/test/Benchmarks/Microsoft.Azure.WebJobs.Script.Tests.Benchmarks.csproj
@@ -1,0 +1,19 @@
+﻿<Project Sdk="Microsoft.NET.Sdk">
+
+  <PropertyGroup>
+    <OutputType>Exe</OutputType>
+    <TargetFramework>netcoreapp3.1</TargetFramework>
+    <AssemblyName>Microsoft.Azure.WebJobs.Script.Tests.Benchmarks</AssemblyName>
+    <RootNamespace>Microsoft.Azure.WebJobs.Script.Tests.Benchmarks</RootNamespace>
+  </PropertyGroup>
+
+  <ItemGroup>
+    <PackageReference Include="BenchmarkDotNet" Version="0.12.1" />
+  </ItemGroup>
+
+  <ItemGroup>
+    <ProjectReference Include="..\..\src\WebJobs.Script.WebHost\WebJobs.Script.WebHost.csproj" />
+    <ProjectReference Include="..\..\src\WebJobs.Script\WebJobs.Script.csproj" />
+  </ItemGroup>
+
+</Project>