--- a/test/Benchmarks/Program.cs
+++ b/test/Benchmarks/Program.cs
@@ -1,0 +1,16 @@
+﻿// Copyright (c) .NET Foundation. All rights reserved.
+// Licensed under the MIT License. See License.txt in the project root for license information.
+
+using System;
+using BenchmarkDotNet.Running;
+
+namespace Microsoft.Azure.WebJobs.Script.Tests.Benchmarks
+{
+    class Program
+    {
+        static void Main(string[] args)
+        {
+            BenchmarkSwitcher.FromAssembly(typeof(Program).Assembly).Run(args);
+        }
+    }
+}