Making a couple minor logging performance improvements, and adding a BenchmarksDotNet project that we can start using for benchmarking. As a sample, I added a benchmark for this change. Here's the before/after results:

| Method |     Mean |    Error |   StdDev |   Median |
|------- |---------:|---------:|---------:|---------:|
| Filter | 473.3 ns | 14.60 ns | 42.13 ns | 458.4 ns |


| Method |     Mean |    Error |   StdDev |   Median |
|------- |---------:|---------:|---------:|---------:|
| Filter | 45.65 ns | 1.301 ns | 3.733 ns | 44.18 ns |

So over a 10x improvement. Now this filtering is a small part of the total request pipeline, so the impact will be minimal, however it's just something I noticed showing up in profiles.

Going forward, we can add more benchmarks for performance critical hotpath methods.